Revert and fix lint
